Add descriptions to tests
