+++ tests_end2end/functional/cases/000_basic/001_simple/description.txt
@@ -1,0 +1,1 @@
+Minimal test case to verify full integration. Ensures that the three main flows (historic, lastdata, and mutable) are working correctly without errors.
