+++ tests_end2end/functional/cases/000_basic/002_streetlight_historic/description.txt
@@ -1,0 +1,1 @@
+Realistic scenario with multiple entity updates to validate correct processing across historic, lastdata, and mutable flows.
