+++ tests_end2end/functional/cases/001_ngsi_types/001_basic_json_types/description.txt
@@ -1,0 +1,1 @@
+Validates correct ingestion and persistence of basic native JSON types: string, number, boolean, null, object, and array.
