+++ tests_end2end/functional/cases/001_ngsi_types/004_text_unrestricted/description.txt
@@ -1,0 +1,1 @@
+Verifies the behavior of the 'Text' type in unrestricted mode, allowing flexible input without strict schema enforcement.
